db created and seeded

+++ migrations/002.do.create_workdays.sql
@@ -1,8 +1,8 @@
 CREATE TABLE "workdays" (
   "id" SERIAL PRIMARY KEY,
   "hours" FLOAT(2) NOT NULL,
-  "downs" TINYINT DEFAULT 0 NOT NULL,
-  "tokes" SMALLINT DEFAULT 0 NOT NULL,
+  "downs" INT DEFAULT 0 NOT NULL,
+  "tokes" INT DEFAULT 0 NOT NULL,
   "date" DATE NOT NULL,
   "notes" TEXT,
   "user_id" INTEGER REFERENCES "user"(id)
